feat: add model for admin
This commit adds the admin models file `adminModels.js`. The models file defines the schema for the admin collection in the MongoDB database. Each document in the admin collection has a unique email and password. The pre-save middleware function hashes the password with bcrypt before saving it to the database, and the comparePassword method is used to compare the candidate password with the hashed password.