refactor: Rename keys and indices for the UserModel schema
This commit updates the keys in the UserModel schema to use underscores instead of camel case and adds database indices to certain fields. Additionally, the 'phoneNumber' field is now 'phone_number' and is no longer a unique field. A new 'verificationToken' field has been added to store a token for user verification.