feat: Add nodemailer email sending functionality to services
This change adds the functionality to send an email using nodemailer library. A function called "sendVerificationEmail" is added to generate and send a verification link email to the users. The transporter object is used to connect to the SMTP server of the email service provider using the credentials from the env file.