feat: Create seller routes for adding and editing products
This commit adds a new file `sellerRouters.js`, which implements two routes for sellers to add and edit products. The `add/products` route is a POST method and requires authentication. The `products/edit/:productId` route is a PUT method and also requires authentication. These routes will be handled by the `addProducts` and `editProduct` functions in the `sellerController` module.