refactor: user registration to fix security vulnerability
Updated the code in the `registerSeller` function to address a security vulnerability related to storing the hashed password in the user object. The modified code separates the steps of updating user data and password hashing, ensuring that the hashed password is securely stored in the database. This commit improves the overall security of the user registration process.